Made a few improvements on webstudio, and add remote machine support
