+++ src/apps/rDSN.WebStudio/app_package/WebStudioCodeDefinition.py
@@ -1,12 +1,7 @@
-import sys
-import os
-import time
-import threading
-
 import WebStudioApp 
+from WebStudioLib import * 
 from dev.python.ServiceApp import *
 THREAD_POOL_DEFAULT = ThreadPoolCode.threadpool_code_register('THREAD_POOL_DEFAULT')
-
 
 class WebStudioService(ServiceApp):
     __server = None
@@ -14,7 +9,6 @@ class WebStudioService(ServiceApp):
     __task = None
 
     def start(self, argv):
-        
         _server_thread= threading.Thread(target=WebStudioApp.start_http_server,args=(argv[1],))
         _server_thread.start()
         return 0
