+++ src/apps/rDSN.WebStudio/app_package/WebStudioUtil.py
@@ -1,0 +1,5 @@
+from WebStudioLib import *
+
+#path helper functions
+def GetWebStudioDirPath():
+    return os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
