+++ src/apps/rDSN.WebStudio/app_package/static/js/service_meta.js
@@ -60,7 +60,7 @@ var vm = new Vue({
                                 {
                                     if(par.primary!='invalid address')
                                     {
-                                        par.membership += 'P: ("' + par.primary + '"),\n ';
+                                        par.membership += 'P: ("' + par.primary + '"), ';
                                         
                                     }
                                     else
