+++ src/apps/rDSN.WebStudio/app_package/static/view/cli.html
@@ -6,7 +6,7 @@
 	</div>
 
 	<div class="page-header">
-        <h1> <font color= #333333><center>Client Local Interface</center></font>
+        <h1> <font color= #333333><center>Command Line Interface</center></font>
 			<small><center><font color= #778899>type 'h' to see profiling commands</font></center></small>
 		</h1>
 	</div>
