--- src/apps/rDSN.WebStudio/app_package/static/view/clusterinfo.html
@@ -1,12 +1,0 @@
-{% extends "static/view/template.html" %}
-{% block body %}
-<pre>
-{{ meta }}
-</pre>
-{% for i in range(replica|length) %}
-<pre>
-{{ replica[i] }}
-</pre>
-{% endfor %}
-{% endblock %}
-
