--- src/apps/rDSN.WebStudio/app_package/static/view/head.html
@@ -1,88 +1,0 @@
-<div class="btn-group">
-	<button class="btn btn-primary dropdown-toggle" @click="collapse">Open task list</button>
-    <button class="btn btn-primary dropdown-toggle" data-toggle="collapse" data-target="#call">Caller/Callee tasks</button>
-	<button class="btn btn-primary dropdown-toggle" data-toggle="collapse" data-target="#linkgraph">Call link graph</button>
-	<button class="btn btn-primary dropdown-toggle" data-toggle="collapse" data-target="#sharer" >Tasks sharing the same threads</button>
-	{% if COMPAREBUTTON=='yes'%}<a href="?task_code={{TASK_CODE}}&ifcompare=yes" class="btn btn-info" role="button">Compare</a>{% endif %}
-</div>
-
-<div id="call" class="collapse">
-	<div class="table-responsive">
-		<table class="table table-striped">
-			<thead>
-				<tr>
-					<th>Callers</th>
-					<th># of calls</th>
-					<th>Callees</th>
-					<th># of calls</th>
-				</tr>
-			</thead>
-			<tbody>
-		
-			</tbody>
-		</table>
-	</div>
-</div>
-
-<style>
-.centered {  
-  position:fixed;
-  z-index: 100;  
-  top:50%;  
-  left:50%;  
-  margin:-100px 0 0 -100px;  
-  width:200px;  
-  height:200px;  
-}  
-.node rect {
-  cursor: move;
-  shape-rendering: crispEdges;
-}
-
-.node text {
-  pointer-events: none;
-  text-shadow: 1px 1px 2px #fff;
-  font-size: 0.8em;
-  font-family: sans-serif;
-}
-
-#tooltip {
-  position: absolute;
-  pointer-events: none;
-  font-size: 0.7em;
-  font-family: sans-serif;
-  padding: 3px;
-  width: auto;
-  height: auto;
-  background-color: #F2F2F2;
-  -webkit-box-shadow: 0px 0px 5px rgba(0, 0, 0, 0.4);
-  -mox-box-shadow: 0px 0px 0px 5px rgba(0, 0, 0, 0.4);
-  box-shadow: 0px 0px 5px rbga(0, 0, 0, 0.4);
-  pointer-events: none;
-}
-
-.value {
-  white-space: pre-line;
-  margin: 0;
-}
-</style>
-<script src="js/sankey_d3.js"></script>
-<script src="js/bihisankey.js"></script>
-<div id="linkgraph" class="collapse">
-    <div id="chart_dep"></div>
-</div>	
-
-<div id="sharer" class="collapse">
-	<div class="table-responsive">
-		<table class="table">
-			<thead>
-				<tr>
-					<th>Sharer</th>
-					<th>Sharer</th>
-				</tr>
-			</thead>
-			<tbody>
-			</tbody>
-		</table>
-	</div>
-</div>
