--- src/apps/rDSN.WebStudio/app_package/static/view/sample.html
@@ -1,136 +1,0 @@
-{% extends "static/view/template.html" %}
-{% block body %}
-<style>
-#wrapper {
-    padding-left: 0;
-    -webkit-transition: all 0.5s ease;
-    -moz-transition: all 0.5s ease;
-    -o-transition: all 0.5s ease;
-    transition: all 0.5s ease;
-}
-
-#wrapper.toggled {
-    padding-left: 400px;
-}
-
-#sidebar-wrapper {
-    z-index: 1000;
-    position: fixed;
-    left: 400px;
-    width: 0;
-    height: 100%;
-    margin-left: -400px;
-    overflow-y: auto;
-    background: #fff;
-    -webkit-transition: all 0.5s ease;
-    -moz-transition: all 0.5s ease;
-    -o-transition: all 0.5s ease;
-    transition: all 0.5s ease;
-}
-
-#wrapper.toggled #sidebar-wrapper {
-    width: 400px;
-}
-
-#page-content-wrapper {
-    width: 100%;
-    position: absolute;
-    padding: 15px;
-}
-
-#wrapper.toggled #page-content-wrapper {
-    position: absolute;
-    margin-right: -400px;
-}
-
-/* Sidebar Styles */
-
-.sidebar-nav {
-    position: absolute;
-    top: 0;
-    width: 400px;
-    margin: 0;
-    padding: 0;
-    list-style: none;
-}
-
-.sidebar-nav li {
-    text-indent: 20px;
-    line-height: 35px;
-}
-
-.sidebar-nav li {
-    display: block;
-    text-decoration: none;
-    color: #DAA520;
-}
-
-.sidebar-nav li:hover {
-    text-decoration: none;
-    color: #7D1935;
-    background: #CBE32D;
-    cursor: pointer;
-}
-
-.sidebar-nav li.active {
-    text-decoration: none;
-    color: #7D1935;
-    background: #A8CD1B;
-}
-
-.sidebar-nav li:focus {
-    text-decoration: none;
-}
-
-.sidebar-nav > .sidebar-brand {
-    height: 65px;
-    font-size: 18px;
-    line-height: 60px;
-    background: #F3FAB6;
-}
-
-.sidebar-nav > .sidebar-brand a {
-    color: #DAA520;
-}
-
-.sidebar-nav > .sidebar-brand a:hover {
-    color: #fff;
-    background: none;
-}
-</style>
-
-
-<div id="app">
-    <div id="wrapper" class="toggled">
-        <div id="sidebar-wrapper">
-            <ul class="sidebar-nav">
-              <li>&nbsp</li>
-              <li class="sidebar-brand"><a href="#">Task List</a></li>
-              <li v-for="task in taskList" :class="[task==currentTask?'active':'']" @click="setcurrentTask(task)">
-                {{task}}
-              </li>
-            </ul>
-        </div>
-        <div id="page-content-wrapper" >
-            <div class="container-fluid">
-                <div class="row">
-                    <div class="page-header">
-                    </div>
-
-                    <div class="page-header">
-                        <h1> <font color= #333333><center>Task Sample</center></font>
-                            <small><center><font color= #778899>Latency distributions of tasks</font></center></small>
-                        </h1>
-                    </div>
-                    {% include "static/view/head.html" %}
-                    <div id="chart"></div>
-                    {% include "static/view/tail.html" %}
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-
-<script src="js/vue.js"></script>
-<script src="js/sample.js"></script>
-{% endblock %}
