--- src/apps/rDSN.WebStudio/app_package/static/view/selection.html
@@ -1,196 +1,0 @@
-{% extends "static/view/template.html" %}
-{% block body %}
-
-<button class="btn btn-primary dropdown-toggle" data-toggle="collapse" data-target="#call">Select counters</button>
-<div class="dropdown">
-<button type="button" class="btn btn-primary dropdown-toggle" data-toggle="dropdown"  data-toggle="tooltip" title="Choose query type">
-    Choose Type ({% if queryType != '' %}{{queryType}}{% else %}sample{% endif %}) <span class="caret"></span></button>
-	<ul class="dropdown-menu" role="menu">
-        <li><a href="?queryType=sample">sample</a></li>
-		<li><a href="?queryType=value">value</a></li>
-	</ul>
-</div>
-
-<script>
-$(function () {
-        $('.list-group.checked-list-box .list-group-item').each(function () {
-
-            // Settings
-            var $widget = $(this),
-            $checkbox = $('<input type="checkbox" class="hidden" />'),
-            color = ($widget.data('color') ? $widget.data('color') : "primary"),
-            style = ($widget.data('style') == "button" ? "btn-" : "list-group-item-"),
-            settings = {
-on: {
-icon: 'glyphicon glyphicon-check'
-},
-off: {
-icon: 'glyphicon glyphicon-unchecked'
-}
-};
-
-$widget.css('cursor', 'pointer')
-$widget.append($checkbox);
-
-// Event Handlers
-$widget.on('click', function () {
-    $checkbox.prop('checked', !$checkbox.is(':checked'));
-    $checkbox.triggerHandler('change');
-    updateDisplay();
-    });
-$checkbox.on('change', function () {
-        updateDisplay();
-        });
-
-
-// Actions
-function updateDisplay() {
-    var isChecked = $checkbox.is(':checked');
-
-    // Set the button's state
-    $widget.data('state', (isChecked) ? "on" : "off");
-
-    // Set the button's icon
-    $widget.find('.state-icon')
-        .removeClass()
-        .addClass('state-icon ' + settings[$widget.data('state')].icon);
-
-    // Update the button's color
-    if (isChecked) {
-        $widget.addClass(style + color + ' active');
-    } else {
-        $widget.removeClass(style + color + ' active');
-    }
-}
-
-// Initialization
-function init() {
-
-    if ($widget.data('checked') == true) {
-        $checkbox.prop('checked', !$checkbox.is(':checked'));
-    }
-
-    updateDisplay();
-
-    // Inject the icon if applicable
-    if ($widget.find('.state-icon').length == 0) {
-        $widget.prepend('<span class="state-icon ' + settings[$widget.data('state')].icon + '"></span>');
-    }
-}
-init();
-});
-
-
-$('#get-checked-data').on('click', function(event) {
-        checkedItemsText = {};
-        checkedItemsIndex = {};
-        counter = 0;
-        event.preventDefault(); 
-        $("#check-list-box li.active").each(function(idx, li) {
-            checkedItemsText[counter] = $(li).text();
-            checkedItemsIndex[counter] = $(li).attr('index');
-            counter++;
-        });
-        basedata = [['x', '1', '2', '3','4','5','6','7','8','9','10', '11', '12', '13','14','15','16','17','18','19'],];
-        for (i=0;i<counter;++i)
-        {
-            basedata.push([checkedItemsText[i],0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0]);
-        }
-        chart = c3.generate({
-            data: {
-                x: 'x',
-                columns: basedata
-            },
-            axis: {
-                x: {
-                    show:false
-                }
-            }
-        });
-        window.setTimeout(function () {
-            updateData(21);
-        }, 3000);
-});
-});
-</script>
-<style>
-/* CSS REQUIRED */
-.state-icon {
-    left: -5px;
-}
-.list-group-item-primary {
-    color: rgb(247, 247, 198);
-    background-color: rgb(66, 139, 202);
-}
-
-/* DEMO ONLY - REMOVES UNWANTED MARGIN */
-.well .list-group {
-    margin-bottom: 0px;
-}
-</style>
-<div class="container">
-    <div id="call" class="collapse">
-        <div class="well" style="max-height: 300px;overflow: auto;">
-            <ul id="check-list-box" class="list-group checked-list-box">
-                {% for app in COUNTER_LIST.items() %}
-                    {% for section in app[1] %}
-                        {% for counter in app[1][section] %}
-                            <li class="list-group-item" index="{{counter['index']}}">{{app[0]}}*{{section}}*{{counter['name']}}</li>
-                        {% endfor %}
-                    {% endfor %}
-                {% endfor %}
-            </ul>
-        </div>
-        <button class="btn btn-primary col-xs-12" id="get-checked-data">Start monitoring!</button>
-    </div>
-</div>
-
-
-<div id="chart"></div>
-<script>
-var basedata = [
-    ['x', '1', '2', '3','4','5','6','7','8','9','10', '11', '12', '13','14','15','16','17','18','19'],
-]
-var latestdata=[['x', 1]]
-var time='';
-var chart = c3.generate({
-    data: {
-        x: 'x',
-        columns: basedata
-    },
-    axis: {
-        x: {
-            show:false
-        }
-    }
-});
-var checkedItemsIndex = {}, checkedItemsText={}, counter = 0;
-
-function updateData(a)
-{
-    $.post("selection.html", {counter_list: JSON.stringify(checkedItemsIndex, null, '\t')}, function(result){
-            latestdata = [['x', a]]
-            parsed_data = JSON.parse(result)
-            for (i=0;i<counter;++i){
-                latestdata.push([checkedItemsText[i],parsed_data.data[i]]);
-            }
-            time = parsed_data.time;
-
-            });
-    chart.flow({
-        columns: latestdata,
-        duration:1000,
-        done:function(){
-            setTimeout(function () {
-                chart.xgrids.add([{value: a, text:time,class:'hoge'}]);
-                updateData(a+1);
-
-            }, 0);}
-    });
-}
-
-
-
-</script>
-{% include "static/view/tail.html" %}
-{% endblock %}
