--- src/apps/rDSN.WebStudio/app_package/static/view/service.html
@@ -1,54 +1,0 @@
-{% extends "static/view/template.html" %}
-{% block body %}
-<style>
-
-.modal.modal-wide .modal-dialog {
-  width: 90%;
-}
-.modal-wide .modal-body {
-  overflow-y: auto;
-}
-.glyphicon:hover {
-    background-color: #ffff00;
-    cursor: pointer;
-}
-</style>
-
-<link href="https://cdn.datatables.net/1.10.9/css/jquery.dataTables.min.css" rel="stylesheet">
-
-<div class="container">
-
-	<div class="page-header">
-	</div>
-
-	<div class="page-header">
-        <h1> <font color= #333333><center>Services</center></font>
-			<small><center><font color= #778899></font></center></small>
-		</h1>
-	</div>
-
-</div>
-<div class="container" style="margin-top:20px;">
-    <div class="row">
-        <table class="table table-striped" id="serviceList">
-        <thead>
-        <tr>
-            <th>Instance</th>
-            <th>Package ID</th>
-            <th>Cluster</th>
-            <th>Service URL</th>
-            <th>State</th>
-            <th>Error</th>
-            <th>Remove</th>
-        </tr>
-        </thead>
-        <tbody>
-        </tbody>
-        </table>
-    </div>
-</div>
-
-
-<script src="js/service.js"></script>
-
-{% endblock %}
