+++ src/apps/rDSN.WebStudio/app_package/static/view/table.html
@@ -14,6 +14,15 @@
         </div>
     </div>
 
+    <div class="row">
+        <div class="col-xs-12">
+            <span>Filter percentage:</span>
+        <select v-model="filterKey">
+            <option selected></option>
+            <option v-for="percent in percentList">{{percent}}</option>
+        </select>
+        </div>
+    </div>
     <table id="table" class="display" cellspacing="0" width="100%">
         <thead>
             <tr>
