--- src/apps/rDSN.WebStudio/app_package/static/view/tail.html
@@ -1,10 +1,0 @@
-
-<div class="btn-group">
-	<button class="btn btn-primary dropdown-toggle" data-toggle="collapse" data-target="#pic" data-toggle="tooltip" title="A help graph for RPC">Show help graph</button>
-</div>
-
-<div id="pic" class="collapse">
-	<center>
-<img src="img/helpgraph.jpg" class="img-rounded" alt="Cinque Terre" width="650" height="425">
-	</center>
-</div>
