+++ src/apps/rDSN.WebStudio/app_package/static/view/template.html
@@ -59,7 +59,7 @@ body {
 		<ul class="nav navbar-nav">
 			<li><a href="main.html" data-toggle="tooltip bottom" title="A overview of the whole system"><span class="glyphicon glyphicon-home"></span> Overview</a></li>
             <li class="dropdown">
-			    <a href="#" class="dropdown-toggle" data-toggle="dropdown"><span class="glyphicon glyphicon-pencil"></span>Program<span class="caret"></span></a>
+			    <a href="#" class="dropdown-toggle" data-toggle="dropdown"><span class="glyphicon glyphicon-pencil"></span>Coding<span class="caret"></span></a>
                 <ul class="dropdown-menu">
 			        <li><a href="#">(None)</a></li>
                 </ul>
@@ -85,7 +85,7 @@ body {
                 </ul>
             </li>
             <li class="dropdown">
-			    <a href="#" class="dropdown-toggle" data-toggle="dropdown"><span class="glyphicon glyphicon-wrench"></span>Run-time Operation<span class="caret"></span></a>
+			    <a href="#" class="dropdown-toggle" data-toggle="dropdown"><span class="glyphicon glyphicon-wrench"></span>Operation<span class="caret"></span></a>
                 <ul class="dropdown-menu">
                     <li><a href="analyzer.html" data-toggle="tooltip right" title="Build your own counter view with counter analyzer!"><span class="glyphicon glyphicon-star"></span>Counter Analyzer</a></li>
                     <li><a href="task_analyzer.html" data-toggle="tooltip right" title="Dig into a specific task to see more detailed information"><span class="glyphicon glyphicon-certificate"></span>Task Analyzer</a></li>
